(BREAKING)Rename procedule module to orderexpansion module
