+++ api/src/main/resources/ProcedureOrder.hbm.xml
@@ -4,7 +4,7 @@
         "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
 
 <hibernate-mapping>
-    <joined-subclass name="org.openmrs.module.procedures.api.model.ProcedureOrder" extends="org.openmrs.Order" table="procedure_order" lazy="false">
+    <joined-subclass name="org.openmrs.module.orderexpansion.api.model.ProcedureOrder" extends="org.openmrs.Order" table="procedure_order" lazy="false">
 
         <key column="order_id" not-null="true" on-delete="cascade" />
 
@@ -23,10 +23,10 @@
         <many-to-one name="frequency" class="org.openmrs.OrderFrequency" />
         <many-to-one name="location" class="org.openmrs.Concept" />
 
-        <many-to-one name="relatedProcedure" class="org.openmrs.module.procedures.api.model.ProcedureOrder"  column="related_procedure_id"/>
+        <many-to-one name="relatedProcedure" class="org.openmrs.module.orderexpansion.api.model.ProcedureOrder"  column="related_procedure_id"/>
         <set name="procedures" table="procedures" inverse="true" cascade="all">
             <key column="procedure_order_id"/>
-            <one-to-many class="org.openmrs.module.procedures.api.model.Procedure"/>
+            <one-to-many class="org.openmrs.module.orderexpansion.api.model.Procedure"/>
         </set>
     </joined-subclass>
 </hibernate-mapping>
