+++ omod/src/main/java/org/openmrs/module/orderexpansion/web/resources/MedicalSupplyDispenseResource.java
@@ -1,10 +1,10 @@
-package org.openmrs.module.procedures.web.resources;
+package org.openmrs.module.orderexpansion.web.resources;
 
 import java.util.Optional;
 
 import org.openmrs.api.context.Context;
-import org.openmrs.module.procedures.api.MedicalSupplyDispenseService;
-import org.openmrs.module.procedures.api.model.MedicalSupplyDispense;
+import org.openmrs.module.orderexpansion.api.MedicalSupplyDispenseService;
+import org.openmrs.module.orderexpansion.api.model.MedicalSupplyDispense;
 import org.openmrs.module.webservices.rest.web.RequestContext;
 import org.openmrs.module.webservices.rest.web.RestConstants;
 import org.openmrs.module.webservices.rest.web.annotation.Resource;
