+++ omod/src/main/java/org/openmrs/module/orderexpansion/web/resources/ProcedureOrderSubclassHandler.java
@@ -1,12 +1,27 @@
-package org.openmrs.module.procedures.web.resources;
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ * <p>
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs.module.orderexpansion.web.resources;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import io.swagger.models.Model;
 import io.swagger.models.ModelImpl;
 import io.swagger.models.properties.IntegerProperty;
+import io.swagger.models.properties.RefProperty;
 import io.swagger.models.properties.StringProperty;
 import org.openmrs.Order;
 import org.openmrs.api.context.Context;
-import org.openmrs.module.procedures.api.model.MedicalSupplyOrder;
+import org.openmrs.module.orderexpansion.api.model.Procedure;
+import org.openmrs.module.orderexpansion.api.model.ProcedureOrder;
+import org.openmrs.module.webservices.docs.swagger.core.property.EnumProperty;
 import org.openmrs.module.webservices.rest.web.RequestContext;
 import org.openmrs.module.webservices.rest.web.annotation.PropertyGetter;
 import org.openmrs.module.webservices.rest.web.annotation.SubClassHandler;
@@ -21,12 +36,12 @@ import org.openmrs.module.webservices.rest.web.resource.impl.DelegatingResourceD
 import org.openmrs.module.webservices.rest.web.resource.impl.DelegatingSubclassHandler;
 import org.openmrs.module.webservices.rest.web.response.ResourceDoesNotSupportOperationException;
 
-@SubClassHandler(supportedClass = MedicalSupplyOrder.class, supportedOpenmrsVersions = { "2.6.* - 9.*" })
-public class MedicalSupplyOrderSubclassHandler extends BaseDelegatingSubclassHandler<Order, MedicalSupplyOrder> implements DelegatingSubclassHandler<Order, MedicalSupplyOrder> {
+@SubClassHandler(supportedClass = ProcedureOrder.class, supportedOpenmrsVersions = { "2.6.* - 9.*" })
+public class ProcedureOrderSubclassHandler extends BaseDelegatingSubclassHandler<Order, ProcedureOrder> implements DelegatingSubclassHandler<Order, ProcedureOrder> {
 	
 	@Override
 	public String getTypeName() {
-		return "medicalsupplyorder";
+		return "procedureorder";
 	}
 	
 	@Override
@@ -35,8 +50,8 @@ public class MedicalSupplyOrderSubclassHandler extends BaseDelegatingSubclassHan
 	}
 	
 	@Override
-	public MedicalSupplyOrder newDelegate() {
-		return new MedicalSupplyOrder();
+	public ProcedureOrder newDelegate() {
+		return new ProcedureOrder();
 	}
 	
 	@Override
@@ -47,7 +62,7 @@ public class MedicalSupplyOrderSubclassHandler extends BaseDelegatingSubclassHan
 	}
 	
 	@PropertyGetter("display")
-	public static String getDisplay(MedicalSupplyOrder delegate) {
+	public static String getDisplay(ProcedureOrder delegate) {
 		OrderResource2_3 orderResource = (OrderResource2_3) Context.getService(RestService.class)
 		        .getResourceBySupportedClass(Order.class);
 		return orderResource.getDisplayString(delegate);
@@ -59,19 +74,29 @@ public class MedicalSupplyOrderSubclassHandler extends BaseDelegatingSubclassHan
 			OrderResource2_3 orderResource = (OrderResource2_3) Context.getService(RestService.class)
 			        .getResourceBySupportedClass(Order.class);
 			DelegatingResourceDescription d = orderResource.getRepresentationDescription(rep);
-			d.addProperty("quantity", Representation.REF);
-			d.addProperty("medicalSuppliesInventoryId");
-			d.addProperty("brandName");
-			d.addProperty("quantityUnits", Representation.REF);
+			d.addProperty("specimenSource", Representation.REF);
+			d.addProperty("laterality");
+			d.addProperty("clinicalHistory");
+			d.addProperty("frequency", Representation.REF);
+			d.addProperty("numberOfRepeats");
+			d.addProperty("specimenType", Representation.REF);
+			d.addProperty("bodySite", Representation.REF);
+			d.addProperty("relatedProcedure", Representation.REF);
+			d.addProperty("procedures", Representation.REF);
 			return d;
 		} else if (rep instanceof FullRepresentation) {
 			OrderResource2_3 orderResource = (OrderResource2_3) Context.getService(RestService.class)
 			        .getResourceBySupportedClass(Order.class);
 			DelegatingResourceDescription d = orderResource.getRepresentationDescription(rep);
-			d.addProperty("quantity", Representation.FULL);
-			d.addProperty("medicalSuppliesInventoryId");
-			d.addProperty("brandName");
-			d.addProperty("quantityUnits", Representation.DEFAULT);
+			d.addProperty("specimenSource", Representation.FULL);
+			d.addProperty("laterality");
+			d.addProperty("clinicalHistory");
+			d.addProperty("frequency", Representation.DEFAULT);
+			d.addProperty("numberOfRepeats");
+			d.addProperty("specimenType", Representation.FULL);
+			d.addProperty("bodySite", Representation.FULL);
+			d.addProperty("relatedProcedure", Representation.FULL);
+			d.addProperty("procedures", Representation.FULL);
 			return d;
 		} else if (rep instanceof CustomRepresentation) { // custom rep
 			return null;
@@ -79,15 +104,33 @@ public class MedicalSupplyOrderSubclassHandler extends BaseDelegatingSubclassHan
 		return null;
 	}
 	
+	@PropertyGetter(value = "procedures")
+	public List<Procedure> getProcedures(ProcedureOrder instance) {
+		try {
+			List<Procedure> procedures = new ArrayList<>(instance.getProcedures());
+			return procedures;
+		}
+		catch (Exception e) {
+			return new ArrayList<>();
+		}
+	}
+	
 	@Override
 	public DelegatingResourceDescription getCreatableProperties() {
 		OrderResource2_3 orderResource = (OrderResource2_3) Context.getService(RestService.class)
 		        .getResourceBySupportedClass(Order.class);
 		DelegatingResourceDescription d = orderResource.getCreatableProperties();
-		d.addProperty("quantity");
-		d.addProperty("medicalSuppliesInventoryId");
-		d.addProperty("brandName");
-		d.addProperty("quantityUnits");
+		d.addProperty("specimenSource");
+		d.addProperty("laterality");
+		d.addProperty("clinicalHistory");
+		d.addProperty("frequency");
+		d.addProperty("numberOfRepeats");
+		d.addProperty("orderType");
+		d.addProperty("bodySite");
+		d.addProperty("specimenType");
+		d.addProperty("commentToFulfiller");
+		d.addProperty("scheduledDate");
+		d.addProperty("relatedProcedure");
 		return d;
 	}
 	
@@ -96,9 +139,16 @@ public class MedicalSupplyOrderSubclassHandler extends BaseDelegatingSubclassHan
 		OrderResource2_3 orderResource = (OrderResource2_3) Context.getService(RestService.class)
 		        .getResourceBySupportedClass(Order.class);
 		ModelImpl orderModel = (ModelImpl) orderResource.getGETModel(rep);
-		orderModel.property("medicalSuppliesInventoryId", new StringProperty()).property("quantity", new IntegerProperty())
-		        .property("brandName", new StringProperty()).property("quantityUnits", new StringProperty());
+		orderModel.property("laterality", new EnumProperty(ProcedureOrder.Laterality.class))
+		        .property("clinicalHistory", new StringProperty()).property("numberOfRepeats", new IntegerProperty());
 		
+		if (rep instanceof DefaultRepresentation) {
+			orderModel.property("specimenSource", new RefProperty("#/definitions/ConceptGetRef")).property("frequency",
+			    new RefProperty("#/definitions/OrderfrequencyGetRef"));
+		} else if (rep instanceof FullRepresentation) {
+			orderModel.property("specimenSource", new RefProperty("#/definitions/ConceptGet")).property("frequency",
+			    new RefProperty("#/definitions/OrderfrequencyGet"));
+		}
 		return orderModel;
 	}
 	
@@ -107,9 +157,11 @@ public class MedicalSupplyOrderSubclassHandler extends BaseDelegatingSubclassHan
 		OrderResource2_3 orderResource = (OrderResource2_3) Context.getService(RestService.class)
 		        .getResourceBySupportedClass(Order.class);
 		ModelImpl orderModel = (ModelImpl) orderResource.getCREATEModel(rep);
-		return orderModel.property("medicalSuppliesInventoryId", new StringProperty().example("uuid"))
-		        .property("quantityUnits", new StringProperty()).property("brandName", new StringProperty())
-		        .property("quantity", new IntegerProperty());
+		return orderModel.property("specimenSource", new StringProperty().example("uuid"))
+		        .property("laterality", new EnumProperty(ProcedureOrder.Laterality.class))
+		        .property("clinicalHistory", new StringProperty())
+		        .property("frequency", new StringProperty().example("uuid"))
+		        .property("numberOfRepeats", new IntegerProperty());
 	}
 	
 }
